twitter/requirements.txt
tweepy==4.6.0
python-dotenv==0.20.0
kafka-python==2.0.2
